% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{args_wo_defval}
\alias{args_wo_defval}
\title{Determine the names of a pairlist without a default value}
\usage{
args_wo_defval(sig)
}
\arguments{
\item{sig}{Pairlist.}
}
\value{
Character vector.
}
\description{
Determine the names of a pairlist without a default value
}
\keyword{internal}

