% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{call_fn}
\alias{call_fn}
\title{Call a function against a given call signature}
\usage{
call_fn(.f)
}
\arguments{
\item{.f}{Function.}
}
\value{
Function of a call (as a language object).
}
\description{
Call a function against a given call signature
}
\keyword{internal}

