% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{checker-object}
\alias{checker-object}
\title{Object checkers}
\arguments{
\item{\dots}{Expressions to validate.}
}
\description{
Object checkers
}
\examples{
row_sums <- firmly(rowSums, vld_matrix(x))
row_sums(matrix(1:6, 2, 3))
\dontrun{
row_sums(mtcars)}

}
\alias{vld_call}
\alias{vld_factor}
\alias{vld_data_frame}
\alias{vld_matrix}
\alias{vld_formula}
\alias{vld_function}
\usage{
vld_call(...)

vld_factor(...)

vld_data_frame(...)

vld_matrix(...)

vld_formula(...)

vld_function(...)
}
