% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{checker-pattern}
\alias{checker-pattern}
\title{Pattern checkers}
\arguments{
\item{\dots}{Expressions to validate.}

\item{pattern}{Regular expression.}

\item{ignore.case}{Should pattern matching ignore case?}

\item{perl}{Should Perl-compatible regular expressions be used?}

\item{prefix, suffix}{String to match.}

\item{na.rm}{Should \code{NA} values be disregarded?}
}
\description{
Pattern checkers
}
\details{
To maintain consistency with \code{\link[=grepl]{grepl()}}, \code{vld_starts_with()} and
\code{vld_ends_with()} coerce to \code{character} before matching.
}
\examples{
ymd <- function(y, m, d) paste(y, m, d, sep = "/")

too_old <- "Not a 21st-century year"
recent_ymd <- firmly(ymd, too_old := vld_grepl("^20[[:digit:]]{2}$", y))

recent_ymd(2017, 01, 01)
\dontrun{
recent_ymd(1999, 01, 01)}

way_too_old <- "Pre-2010 year is too old"
more_recent_ymd <- firmly(ymd, way_too_old := vld_starts_with("201", y))

more_recent_ymd(2017, 01, 01)
\dontrun{
more_recent_ymd(2001, 01, 01)}

}
\seealso{
\code{\link[=grepl]{grepl()}}, \code{\link[=startsWith]{startsWith()}}, \code{\link[=endsWith]{endsWith()}}
}
\alias{vld_grepl}
\alias{vld_starts_with}
\alias{vld_ends_with}
\usage{
vld_grepl(pattern, ..., ignore.case = FALSE, perl = FALSE)

vld_starts_with(prefix, ..., na.rm = FALSE)

vld_ends_with(suffix, ..., na.rm = FALSE)
}
