% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{checker-property}
\alias{checker-property}
\title{Property checkers}
\arguments{
\item{\dots}{Expressions to validate.}

\item{nm, nms}{Name(s).}

\item{n}{Length of vector or list.}

\item{which}{Object attribute(s).}

\item{what}{Class name.}
}
\description{
Property checkers
}
\seealso{
\code{\link[=vld_null]{vld_null()}}, \code{\link[=vld_not_null]{vld_not_null()}}
}
\alias{vld_not_empty}
\alias{vld_singleton}
\alias{vld_not_na}
\alias{vld_without_na}
\alias{vld_sorted}
\alias{vld_named}
\alias{vld_has_name}
\alias{vld_has_names}
\alias{vld_has_length}
\alias{vld_has_attr}
\alias{vld_has_attrs}
\alias{vld_inherits}
\usage{

vld_not_empty(...)

vld_singleton(...)

vld_not_na(...)

vld_without_na(...)

vld_sorted(...)

vld_named(...)

vld_has_name(nm, ...)

vld_has_names(nms, ...)

vld_has_length(n, ...)

vld_has_attr(which, ...)

vld_has_attrs(which, ...)

vld_inherits(what, ...)
}
