% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{checker-relational}
\alias{checker-relational}
\title{Relational checkers}
\arguments{
\item{\dots}{Expressions to validate.}

\item{to}{Reference object.}

\item{lwr, upr}{Lower/upper bound.}

\item{na.rm}{Should \code{NA} values be disregarded?}
}
\description{
Relational checkers
}
\examples{
f <- function(x, y) log(y - x) / log(x)
foo <- firmly(f, vld_gt(0, x - 1, "y not greater than x" := y - x))
foo(2, 4)
\dontrun{
foo(1, 2)
foo(2, 2)}

}
\seealso{
\code{\link[=all.equal]{all.equal()}}, \code{\link[=identical]{identical()}}
}
\alias{vld_identical}
\alias{vld_not_identical}
\alias{vld_equal}
\alias{vld_not_equal}
\alias{vld_equivalent}
\alias{vld_not_equivalent}
\alias{vld_gt}
\alias{vld_lt}
\alias{vld_gte}
\alias{vld_lte}
\usage{
vld_identical(to, ...)

vld_not_identical(to, ...)

vld_equal(to, ...)

vld_not_equal(to, ...)

vld_equivalent(to, ...)

vld_not_equivalent(to, ...)

vld_gt(lwr, ..., na.rm = FALSE)

vld_lt(upr, ..., na.rm = FALSE)

vld_gte(lwr, ..., na.rm = FALSE)

vld_lte(upr, ..., na.rm = FALSE)
}
