% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{checker-scalar-type}
\alias{checker-scalar-type}
\title{Scalar type checkers}
\arguments{
\item{\dots}{Expressions to validate.}
}
\description{
Scalar type checkers
}
\seealso{
\code{\link[=vld_singleton]{vld_singleton()}}
}
\alias{vld_number}
\alias{vld_boolean}
\alias{vld_string}
\usage{
vld_number(...)

vld_boolean(...)

vld_string(...)
}
