% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{checkers-bare-type}
\alias{checkers-bare-type}
\title{Bare type checkers}
\arguments{
\item{...}{Expressions to check.}
}
\value{
Check formula.
}
\description{
Bare type checkers
}
\seealso{
\link[purrr:bare-type-predicates]{Bare type predicates} (purrr)

Other type checkers: \code{\link{checkers-misc}},
  \code{\link{checkers-scalar-type}},
  \code{\link{checkers-type}}
}
\alias{vld_bare_atomic}
\alias{vld_bare_character}
\alias{vld_bare_double}
\alias{vld_bare_integer}
\alias{vld_bare_list}
\alias{vld_bare_logical}
\alias{vld_bare_numeric}
\alias{vld_bare_vector}
\usage{
vld_bare_atomic(...)

vld_bare_character(...)

vld_bare_double(...)

vld_bare_integer(...)

vld_bare_list(...)

vld_bare_logical(...)

vld_bare_numeric(...)

vld_bare_vector(...)
}
