% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{checkers-misc}
\alias{checkers-misc}
\title{Miscellaneous type checkers}
\arguments{
\item{...}{Expressions to check.}
}
\value{
Check formula.
}
\description{
Miscellaneous type checkers
}
\seealso{
Other type checkers: \code{\link{checkers-bare-type}},
  \code{\link{checkers-scalar-type}},
  \code{\link{checkers-type}}
}
\alias{vld_empty}
\alias{vld_formula}
\usage{
vld_empty(...)

vld_formula(...)
}
