% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{checkers-scalar}
\alias{checkers-scalar}
\title{Scalar type checkers}
\arguments{
\item{...}{Expressions to check.}
}
\value{
Check formula.
}
\description{
Scalar type checkers
}
\alias{vld_scalar_atomic}
\alias{vld_scalar_character}
\alias{vld_scalar_double}
\alias{vld_scalar_integer}
\alias{vld_scalar_list}
\alias{vld_scalar_logical}
\alias{vld_scalar_numeric}
\alias{vld_scalar_vector}
\usage{
vld_scalar_atomic(...)

vld_scalar_character(...)

vld_scalar_double(...)

vld_scalar_integer(...)

vld_scalar_list(...)

vld_scalar_logical(...)

vld_scalar_numeric(...)

vld_scalar_vector(...)
}
