% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{checkers-type}
\alias{checkers-type}
\title{Type checkers}
\arguments{
\item{...}{Expressions to check.}
}
\value{
Check formula.
}
\description{
Type checkers
}
\seealso{
Other type checkers: \code{\link{checkers-bare-type}},
  \code{\link{checkers-misc}},
  \code{\link{checkers-scalar-type}}
}
\alias{vld_atomic}
\alias{vld_character}
\alias{vld_double}
\alias{vld_function}
\alias{vld_integer}
\alias{vld_list}
\alias{vld_logical}
\alias{vld_null}
\alias{vld_numeric}
\alias{vld_vector}
\usage{
vld_atomic(...)

vld_character(...)

vld_double(...)

vld_function(...)

vld_integer(...)

vld_list(...)

vld_logical(...)

vld_null(...)

vld_numeric(...)

vld_vector(...)
}
