% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{checkers}
\alias{checkers}
\title{Make local check formulae}
\arguments{
\item{...}{Expressions of function arguments to check.}
}
\value{
Check formula.
}
\description{
Check makers derived from purrr predicate functions.
}
\alias{vld_scalar_vector}
\alias{vld_bare_integer}
\alias{vld_bare_vector}
\alias{vld_bare_atomic}
\alias{vld_empty}
\alias{vld_formula}
\alias{vld_bare_character}
\alias{vld_character}
\alias{vld_bare_numeric}
\alias{vld_numeric}
\alias{vld_logical}
\alias{vld_scalar_numeric}
\alias{vld_integer}
\alias{vld_vector}
\alias{vld_scalar_logical}
\alias{vld_atomic}
\alias{vld_function}
\alias{vld_bare_list}
\alias{vld_scalar_double}
\alias{vld_list}
\alias{vld_bare_double}
\alias{vld_null}
\alias{vld_scalar_list}
\alias{vld_double}
\alias{vld_scalar_atomic}
\alias{vld_bare_logical}
\alias{vld_scalar_character}
\alias{vld_scalar_integer}
\usage{
vld_scalar_vector(...)

vld_bare_integer(...)

vld_bare_vector(...)

vld_bare_atomic(...)

vld_empty(...)

vld_formula(...)

vld_bare_character(...)

vld_character(...)

vld_bare_numeric(...)

vld_numeric(...)

vld_logical(...)

vld_scalar_numeric(...)

vld_integer(...)

vld_vector(...)

vld_scalar_logical(...)

vld_atomic(...)

vld_function(...)

vld_bare_list(...)

vld_scalar_double(...)

vld_list(...)

vld_bare_double(...)

vld_null(...)

vld_scalar_list(...)

vld_double(...)

vld_scalar_atomic(...)

vld_bare_logical(...)

vld_scalar_character(...)

vld_scalar_integer(...)
}
