% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checklist}
\alias{checklist}
\alias{is_check_formula}
\alias{is_checklist}
\title{Is a checklist valid?}
\usage{
is_checklist(x)

is_check_formula(x)
}
\arguments{
\item{x}{R object.}
}
\value{
\code{TRUE} or \code{FALSE}, according to whether \code{x} is a valid
  checklist.
}
\description{
A valid checklist is a list of formulae \code{q ~ p}, where \code{p} is a
function (or a one-sided formulaic shorthand thereof, see
\code{\link[purrr]{as_function}()}), and \code{q} is a "check item" for
\code{p}; a check item is either empty, a string, or a list of formulae of
the form \code{~<expr>} or \code{<string> ~ <expr>}. (\code{p} is to be
called on \code{<expr>}.)
}
\details{
\code{is_checklist()} is used to check the validity of the checks
  passed to \code{\link{strictly}()}, either via \code{...} or the argument
  \code{.checklist}.
}

