% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clone_env}
\alias{clone_env}
\title{Clone an environment}
\usage{
clone_env(env, parent = parent.env(env))
}
\arguments{
\item{env, parent}{Environments.}
}
\value{
Clone of the environment \code{x}.
}
\description{
Clone an environment
}
\details{
Taken from the section
  \href{http://adv-r.had.co.nz/dsl.html#latex}{LaTeX} of the book
  \emph{Advanced R}, by Hadley Wickham.
}
\keyword{internal}

