% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{comparisons-boolean}
\alias{comparisons-boolean}
\title{Boolean comparison predicates}
\arguments{
\item{\dots}{Expressions to validate}

\item{.ref}{Comparison object.}
}
\description{
Boolean comparison predicates
}
\alias{vld_all_map}
\alias{vld_any_map}
\usage{
vld_all_map(...)

vld_any_map(...)
}
