% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{comparisons-pattern}
\alias{comparisons-pattern}
\title{Pattern comparison predicates}
\arguments{
\item{\dots}{Expressions to validate}

\item{.ref}{Comparison object.}
}
\description{
Pattern comparison predicates
}
\alias{vld_grepl}
\usage{
vld_grepl(...)
}
