% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{comparisons-property}
\alias{comparisons-property}
\title{Property comparison predicates}
\arguments{
\item{\dots}{Expressions to validate}

\item{.ref}{Comparison object.}
}
\description{
Property comparison predicates
}
\alias{vld_has_name}
\alias{vld_has_names}
\alias{vld_has_length}
\alias{vld_has_attr}
\alias{vld_has_attrs}
\alias{vld_inherits}
\usage{
vld_has_name(.ref, ...)

vld_has_names(.ref, ...)

vld_has_length(.ref, ...)

vld_has_attr(.ref, ...)

vld_has_attrs(.ref, ...)

vld_inherits(.ref, ...)
}
