% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{comparisons-relation}
\alias{comparisons-relation}
\title{Relation predicates}
\arguments{
\item{\dots}{Expressions to validate}

\item{.ref}{Comparison object.}
}
\description{
Relation predicates
}
\alias{vld_identical}
\alias{vld_not_identical}
\alias{vld_equal}
\alias{vld_not_equal}
\alias{vld_all_equal}
\alias{vld_not_all_equal}
\alias{vld_gt}
\alias{vld_lt}
\alias{vld_gte}
\alias{vld_lte}
\usage{
vld_identical(...)

vld_not_identical(...)

vld_equal(...)

vld_not_equal(...)

vld_all_equal(...)

vld_not_all_equal(...)

vld_gt(...)

vld_lt(...)

vld_gte(...)

vld_lte(...)
}
