% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{comparisons-sets}
\alias{comparisons-sets}
\title{Set comparison predicates}
\arguments{
\item{\dots}{Expressions to validate}

\item{.ref}{Comparison object.}
}
\description{
Set comparison predicates
}
\alias{vld_in}
\alias{vld_not_in}
\alias{vld_contains}
\alias{vld_doesnt_contain}
\alias{vld_contained_in}
\alias{vld_not_contained_in}
\alias{vld_intersects}
\alias{vld_doesnt_intersect}
\alias{vld_setequal}
\usage{
vld_in(...)

vld_not_in(...)

vld_contains(...)

vld_doesnt_contain(...)

vld_contained_in(...)

vld_not_contained_in(...)

vld_intersects(...)

vld_doesnt_intersect(...)

vld_setequal(...)
}
