% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{condition}
\alias{condition}
\title{Create a condition object}
\usage{
condition(subclass, message, call = sys.call(-1), ...)
}
\arguments{
\item{subclass}{Subclass of \code{"condition"} (string).}

\item{message}{Message (string).}

\item{call}{Call object.}

\item{...}{Additional arguments to pass to}
}
\value{
Condition object (class \code{"condition"}).
}
\description{
Create a condition object
}
\details{
Taken from the section
  \href{http://adv-r.had.co.nz/Exceptions-Debugging.html#condition-handling}{Condition
   handling} of the book \emph{Advanced R}, by Hadley Wickham.
}
\keyword{internal}

