% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparse.R
\name{deparse_collapse}
\alias{deparse_collapse}
\title{Deparse a language object as a single string}
\usage{
deparse_collapse(x)
}
\arguments{
\item{x}{Language object.}
}
\value{
String.
}
\description{
Deparse a language object as a single string
}
\details{
Used in validation check template, must therefore be exported.
}
\keyword{internal}

