% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enumerate_many}
\alias{enumerate_many}
\title{Enumerate a character vector as a string}
\usage{
enumerate_many(x, many = 2L)
}
\arguments{
\item{x}{Character vector.}

\item{many}{Integer. How many is too many?}
}
\value{
String.
}
\description{
Enumerate a character vector as a string
}
\details{
Used in validation check template, must therefore be exported.
}
\keyword{internal}

