% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strictly.R
\name{find_data.environment}
\alias{find_data.environment}
\title{Method for \code{lazyeval::find_data}}
\usage{
find_data.environment(x)
}
\description{
Method for \code{lazyeval::find_data}
}
\keyword{internal}

