% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_valid_checklist}
\alias{is_valid_checklist}
\title{Is a checklist valid?}
\usage{
is_valid_checklist(...)
}
\arguments{
\item{...}{R objects.}
}
\value{
\code{TRUE} or \code{FALSE}, according to whether \code{list(...)} is
  a valid checklist.
}
\description{
A valid checklist is either a list of formulas or a list thereof. An empty
list is considered valid.
}
\details{
\code{is_valid_checklist()} is used to check the validity of the
  checklist passed to \code{strictly()}.
}

