% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_valid_cond}
\alias{is_valid_cond}
\title{Is a condition valid?}
\usage{
is_valid_cond(x)
}
\arguments{
\item{x}{R object.}
}
\value{
\code{TRUE} or \code{FALSE}, according to whether \code{x} is
  a valid value for \code{.cond} in \code{strictly()}.
}
\description{
Is a condition valid?
}
\details{
\code{is_valid_cond()} is used to check the validity of the
  condition argument (\code{.cond}) of \code{strictly()}.
}
\keyword{internal}

