% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{localize_check}
\alias{localize_check}
\title{Localize a global check formula}
\usage{
localize_check(chk)
}
\arguments{
\item{chk}{Global check formula.}
}
\value{
Local check formula.
}
\description{
Localize a global check formula
}
\examples{
\dontrun{
is_number <- localize_check("Not a number" ~ purrr::is_scalar_numeric)
secant <- function(f, x, dx) (f(x + dx) - f(x)) / dx
secant <- strictly(secant, is_number(x, dx))
}
}

