% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{make_check}
\alias{make_check}
\title{Customize check functions}
\usage{
make_check(msg, p)
}
\arguments{
\item{msg}{String.}

\item{p}{Predicate function or lambda expression.}
}
\value{
Formula
}
\description{
Customize check functions
}
\examples{
\dontrun{
is_number <- make_check(purrr::is_scalar_numeric, "Not a number")
secant <- function(f, x, dx) (f(x + dx) - f(x)) / dx
secant <- strictly(secant, is_number(x, dx))
}
}

