% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{misc-checkers}
\alias{misc-checkers}
\title{Miscellaneous type checkers}
\arguments{
\item{...}{One-sided formulae that specify the expressions to check.}
}
\value{
Check formula of local scope.
}
\description{
These functions make check formulae of local scope based on the
correspondingly named predicates \code{\link[purrr]{is_empty}()},
\code{\link[purrr]{is_formula}()} from the \pkg{purrr} package.
}
\details{
Each \code{vld_*()} is a function of class \code{"check_maker"}
  generated by \code{\link{localize}()}.
}
\examples{
\dontrun{

f <- function(x, y) "Pass"

# Impose a check on x: ensure it's a formula
f_strict <- strictly(f, vld_formula(~x))
f_strict(z ~ a + b, 0)  # "Pass"
f_strict(0, 0)          # Error: "Not formula: x"

# Impose the check, "x is a subset of y"
f_stricter <- strictly(f, vld_empty(~ setdiff(x, y)))
f_stricter(letters[1:3], letters)  # "Pass"
f_stricter(letters, letters[1:3])  # Error: "Not empty: setdiff(x, y)"
}
}
\seealso{
\code{\link[purrr]{is_empty}}, \code{\link[purrr]{is_formula}}
  (\pkg{purrr})

Other type checkers: \code{\link{bare-type-checkers}},
  \code{\link{scalar-type-checkers}},
  \code{\link{type-checkers}}
}
\alias{vld_empty}
\alias{vld_formula}
\usage{
vld_empty(...)

vld_formula(...)
}
