% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strictly.R
\name{nomen}
\alias{nomen}
\title{Represent non-dot arguments by name and symbol}
\usage{
nomen(sig)
}
\arguments{
\item{sig}{Pairlist.}
}
\value{
List with components \code{"nm"} (character), \code{"symb"} (symbol),
  \code{"wo_value"} (logical).
}
\description{
Represent non-dot arguments by name and symbol
}
\keyword{internal}

