% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null-default}
\alias{\%||\%}
\alias{null-default}
\title{Default value for \code{NULL}.}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{R objects.}
}
\description{
Default value for \code{NULL}.
}
\details{
Taken from the section
  \href{http://adv-r.had.co.nz/Functions.html#special-calls}{Special calls}
  of the book \emph{Advanced R}, by Hadley Wickham.
}
\keyword{internal}

