% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates-object}
\alias{predicates-object}
\title{Object predicates}
\arguments{
\item{x}{Object to test.}
}
\description{
Object predicates
}
\examples{
row_sums <- firmly(rowSums, chk_matrix(x))
row_sums(matrix(1:6, 2, 3))

## Meaningless to sum across rows when column units differ
\dontrun{
row_sums(mtcars)}

}
\seealso{
\link[=predicates-type]{Type predicates}
}
\alias{chk_call}
\alias{chk_factor}
\alias{chk_data_frame}
\alias{chk_matrix}
\alias{chk_formula}
\alias{chk_function}
\usage{
chk_call(x)

chk_factor(x)

chk_data_frame(x)

chk_matrix(x)

chk_formula(x)

chk_function(x)
}
