% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates-pattern}
\alias{predicates-pattern}
\title{Pattern predicates}
\arguments{
\item{x}{Object to test.}

\item{pattern}{Regular expression.}

\item{ignore.case}{Should pattern matching ignore case?}

\item{perl}{Should Perl-compatible regular expressions be used?}

\item{prefix, suffix}{String to match.}

\item{na.rm}{Should \code{NA} values be disregarded?}
}
\description{
Pattern predicates
}
\details{
To maintain consistency with \code{\link[=grepl]{grepl()}}, \code{chk_starts_with()} and
\code{chk_ends_with()} coerce to \code{character} before matching.
}
\examples{
ymd <- function(y, m, d) paste(y, m, d, sep = "/")

too_old <- "Not a 21st-century year"
recent_ymd <- firmly(ymd, !! too_old := chk_grepl("^20[[:digit:]]{2}$", y))

recent_ymd(2017, 01, 01)
\dontrun{
recent_ymd(1999, 01, 01)}

way_too_old <- "Pre-2010 year is too old"
more_recent_ymd <- firmly(ymd, !! way_too_old := chk_starts_with("201", y))

more_recent_ymd(2017, 01, 01)
\dontrun{
more_recent_ymd(2001, 01, 01)}

}
\seealso{
\code{\link[=grepl]{grepl()}}, \code{\link[=startsWith]{startsWith()}}, \code{\link[=endsWith]{endsWith()}}
}
\alias{chk_grepl}
\alias{chk_starts_with}
\alias{chk_ends_with}
\usage{
chk_grepl(x, pattern, ignore.case = FALSE, perl = FALSE)

chk_starts_with(x, prefix, na.rm = FALSE)

chk_ends_with(x, suffix, na.rm = FALSE)
}
