% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates-relation}
\alias{predicates-relation}
\title{Relation predicates}
\arguments{
\item{\dots}{Expressions to validate}
}
\description{
Relation predicates
}
\alias{vld_identical}
\alias{vld_not_identical}
\alias{vld_equal}
\alias{vld_not_equal}
\alias{vld_all_equal}
\alias{vld_not_all_equal}
\alias{vld_gt}
\alias{vld_lt}
\alias{vld_gte}
\alias{vld_lte}
\usage{
vld_identical(y, ...)

vld_not_identical(y, ...)

vld_equal(y, ...)

vld_not_equal(y, ...)

vld_all_equal(target, ...)

vld_not_all_equal(target, ...)

vld_gt(lwr, ...)

vld_lt(upr, ...)

vld_gte(lwr, ...)

vld_lte(upr, ...)
}
