% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates-set}
\alias{predicates-set}
\title{Set predicates}
\arguments{
\item{x}{Object to test.}

\item{set}{Reference set (as a vector).}
}
\description{
Set predicates
}
\examples{
s3methods <- function(x) {
  methods <- attr(methods(class = class(x)), "info")
  with(methods, generic[!isS4])
}
foo <- fasten(
  vld_include("predict", s3methods(object))
)(
  function(object, data) {
    pred <- predict(object, data)
    "Do something with prediction"
  }
)

mdl <- lm(mpg ~ wt, mtcars[1:10, ])
data <- mtcars[11:12, ]

foo(mdl, data)
\dontrun{
foo(NULL, data)}

}
\seealso{
\link[=setequal]{Set operations},
\link[=predicates-property]{Property predicates}
}
\alias{vld_in}
\alias{vld_not_in}
\alias{vld_include}
\alias{vld_exclude}
\alias{vld_within}
\alias{vld_intersect}
\alias{vld_avoid}
\alias{vld_setequal}
\usage{
vld_in(x, set)

vld_not_in(x, set)

vld_include(x, set)

vld_exclude(x, set)

vld_within(x, set)

vld_intersect(x, set)

vld_avoid(x, set)

vld_setequal(x, set)
}
