% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates-sets}
\alias{predicates-sets}
\title{Set comparison predicates}
\arguments{
\item{\dots}{Expressions to validate}
}
\description{
Set comparison predicates
}
\alias{vld_in}
\alias{vld_not_in}
\alias{vld_contains}
\alias{vld_doesnt_contain}
\alias{vld_contained_in}
\alias{vld_not_contained_in}
\alias{vld_intersects}
\alias{vld_doesnt_intersect}
\alias{vld_setequal}
\usage{
vld_in(set, ...)

vld_not_in(set, ...)

vld_contains(set, ...)

vld_doesnt_contain(set, ...)

vld_contained_in(set, ...)

vld_not_contained_in(set, ...)

vld_intersects(set, ...)

vld_doesnt_intersect(set, ...)

vld_setequal(set, ...)
}
