% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates-type}
\alias{predicates-type}
\title{Type predicates}
\arguments{
\item{\dots}{Expressions to validate}
}
\description{
Type predicates
}
\alias{vld_closure}
\alias{vld_numerical}
\alias{vld_null}
\alias{vld_symbol}
\alias{vld_pairlist}
\alias{vld_environment}
\alias{vld_language}
\alias{vld_list}
\alias{vld_logical}
\alias{vld_integer}
\alias{vld_double}
\alias{vld_complex}
\alias{vld_character}
\alias{vld_raw}
\usage{
vld_closure(...)

vld_numerical(...)

vld_null(...)

vld_symbol(...)

vld_pairlist(...)

vld_environment(...)

vld_language(...)

vld_list(...)

vld_logical(...)

vld_integer(...)

vld_double(...)

vld_complex(...)

vld_character(...)

vld_raw(...)
}
