% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promises.R
\name{promises}
\alias{promises}
\title{Capture function-call arguments as promises}
\usage{
promises(.call, .sig, .env)
}
\arguments{
\item{.call}{Function call (language).}

\item{.sig}{Function argument signature (pairlist).}

\item{.env}{Environment in which the function-call arguments are to be
(lazily) evaluated.}
}
\value{
Environment in which the function-call arguments (and no other
  expressions) are bound as promises.
}
\description{
Capture function-call arguments as promises
}
\details{
This is essentially equivalent to using
  \code{\link[base]{delayedAssign}()} in a for-loop to bind promises, but at
  roughly twice the speed and half the code.
}
\keyword{internal}

