% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_check}
\alias{remove_check}
\title{Remove a check from a strict closure}
\usage{
remove_check(..f, which)
}
\arguments{
\item{..f}{Strict closure, i.e., function of class \code{"strict_closure"}.}

\item{which}{Logical or numeric vector by which to subset
\code{sc_check(.f)}.}
}
\value{
Strict closure.
}
\description{
Remove a check from a strict closure
}

