% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repn_args}
\alias{repn_args}
\title{Represent non-dot arguments by name and symbol}
\usage{
repn_args(sig)
}
\arguments{
\item{sig}{Pairlist.}
}
\value{
List with components \code{"nm"} (character), \code{"symb"} (symbol),
  \code{"has_val"} (logical).
}
\description{
Represent non-dot arguments by name and symbol
}
\keyword{internal}

