% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{sc-attributes}
\alias{is_strict_closure}
\alias{sc-attributes}
\alias{sc_arg_req}
\alias{sc_check}
\alias{sc_core}
\alias{sc_logical_void}
\title{Get attributes of a strict closure}
\usage{
is_strict_closure(x)

sc_core(.f)

sc_check(.f)

sc_arg_req(.f)

sc_logical_void(.f)
}
\arguments{
\item{x}{R object.}

\item{.f}{Interpreted function, i.e., closure (not a primitive function).}
}
\value{
\code{sc_core()} returns a language object; \code{sc_check()} returns
  a list of checks, where each check is given as a list consisting of a call
  (language object) and its string representation; \code{sc_arg_req()}
  returns a character vector. \code{NULL} is returned whenever an attribute
  is absent.
}
\description{
A strict closure—a value of the function \code{\link{strictly}()}—stores
attributes of the function it "strictifies." The \code{sc_*()} functions
extract these attributes:
\itemize{
  \item \code{sc_core()} gets the original function body.
  \item \code{sc_check()} gets the checks, as calls to predicate functions.
  \item \code{sc_arg_req()} gets the required arguments, if their absence is
    to be checked.
  \item \code{sc_logical_void()} gets the interpretation of a
    \code{logical(0)} predicate value (which can be either \code{logical(0)},
    \code{TRUE}, or \code{FALSE})
}
\code{is_strict_closure()} checks whether an object is a closure of (S3)
class \code{"strict_closure"}.
}
\examples{
# Probe the attributes of `strictly()` itself, as a strict closure

is_strict_closure(strictly)

sc_core(strictly)
sc_check(strictly)
sc_arg_req(strictly)
}
\seealso{
\code{\link{strictly}}
}

