% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope.R
\name{scope}
\alias{scope}
\alias{localize}
\alias{globalize}
\title{Convert the scope of a check formula}
\usage{
localize(chk)

globalize(chkr)
}
\arguments{
\item{chk}{Check formula of global scope with a custom error message, i.e., a
formula of the form \code{<string> ~ <predicate>}, cf. \link{strictly}.}

\item{chkr}{Function of class \code{"check_maker"}, i.e., created by
\code{localize()}.}
}
\description{
\code{localize()} converts a check formula of global scope into a
function that generates a check formulae of local scope;
\code{globalize()} takes such a function and returns the underlying
check formula (of global scope). These operations are mutually invertible.
}
\examples{
\dontrun{

is_positive <- localize("Not positive" ~ {. > 0})
is_positive(x, x - y)
#> list("Not positive: `x`" ~ x, "Not positive: `x - y`" ~ x - y) ~ {. > 0}

globalize(is_positive)
#> "Not positive" ~ {. > 0}
}

}
\seealso{
\link{strictly} explains the notion of "scope" for check formulae.
}
