% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monads.R
\name{setup_monad}
\alias{setup_monad}
\title{Provide ingredients for making a monad}
\usage{
setup_monad(type, data, metadata, env = parent.frame())
}
\arguments{
\item{type}{Name of monad type (string).}

\item{data, metadata}{Name of data, resp. metadata, fields of monadic object
(string).}

\item{env}{Environment to which to assign monad-structure functions.}
}
\value{
Invisibly returns arguments.
}
\description{
A monad of augmented data records ordinary data (e.g., a numerical value),
together with metadata (e.g., a log message). \code{setup_monad()} sets up
the ingredients for making a monad, named \code{<type>} (string), of
augmented data:
\itemize{
  \item \code{<type>()}: Creates a monadic object from an ordinary one.
  \item \code{<type>_function()}: Designates a function as one with monadic
    output.
  \item \code{is_<type>()}: Is an object a monad?
  \item \code{is_<type>_function()}: Does a function have monadic output?
  \item \code{<type>_unit()}: Unit for monad.
}
(To finish monad setup, one needs to provide a "bind" method for the generic
function \code{bind()}.)
}
\examples{
setup_monad("writer", data = "value", metadata = "log")
writer(value = 3, log = "Three")
writer_unit(3)
}
\keyword{internal}

