% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strictly.R
\name{strict_closure}
\alias{strict_closure}
\title{Create an object of class "strict_closure"}
\usage{
strict_closure(x, ...)
}
\arguments{
\item{x}{R object.}
}
\description{
Create an object of class "strict_closure"
}
\keyword{internal}

