% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unpack}
\alias{unpack}
\title{Unpack a list of a list}
\usage{
unpack(x)
}
\arguments{
\item{x}{List.}
}
\value{
\code{x[[1L]]}, if \code{x} is a list of a list, otherwise \code{x}.
}
\description{
Unpack a list of a list
}
\keyword{internal}

