% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valaddin.R
\docType{package}
\name{valaddin-package}
\alias{valaddin}
\alias{valaddin-package}
\title{valaddin: Functional Input Validation}
\description{
A set of basic tools to transform functions into functions with
input validation checks, in a manner suitable for both programmatic and
interactive use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/egnha/valaddin}
  \item Report bugs at \url{https://github.com/egnha/valaddin/issues}
}

}
\author{
\strong{Maintainer}: Eugene Ha \email{eha@posteo.de}

}
\keyword{internal}
