% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valaddin.R
\docType{package}
\name{valaddin}
\alias{valaddin}
\alias{valaddin-package}
\title{valaddin: Functional Argument Validation}
\description{
Enable input validation for your function, \emph{post factum}. This package
provides a set of basic tools to add argument checks in a functional manner.
}

