% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{validate}
\alias{validate}
\alias{validify}
\title{Validate objects}
\arguments{
\item{.}{Object to validate.}

\item{...}{Input validation checks.}

\item{error_class}{Character vector of the error subclass to signal if
validation fails. If \code{NULL} (the default), the error subclass is
\code{objectValidationError}.}
}
\description{
Validate objects
}
\examples{
# All assertions valid: data frame returned (invisibly)
validate(mtcars,
         is.data.frame,
         chk_all_map(is.numeric),
         chk_gt(10, nrow(.)),
         chk_has_names(c("mpg", "cyl")))

# Some assertions invalid: diagnostic error raised
\dontrun{
validate(mtcars,
         is.matrix,
         chk_all_map(is.numeric),
         chk_gt(1000, nrow(.)),
         chk_has_name("cylinders"))}

}
\usage{
validate(., ..., error_class = NULL)

validify(..., error_class = NULL)
}
