% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collector.R
\name{vld}
\alias{vld}
\title{Collect input validation checks}
\usage{
vld(...)
}
\arguments{
\item{...}{Input-validation checks as error message-check definitions (cf.
\code{\link[rlang:dots_definitions]{rlang::dots_definitions()}}).}
}
\value{
List of pairs of quosures, where each pair has components \code{msg}
(error message) and \code{chk} (input validation check).
}
\description{
Collect input validation checks
}
\details{
The order of the input validation checks is not necessarily
preserved.
}
\examples{
## Use vld() to package input validation checks (which may be nested)
chks <- local({
  zero <- 0
  nonneg <- vld(
    {isTRUE(. >= zero)} ~
      vld(a - b, "b - c is negative ({b - c})" := b - c)
  )
  vld("Not numeric" := is.numeric, !!! nonneg)
})

## Enforce that a, b, c are numeric and a > b > c
f <- firmly(function(a, b, c) list(a, b, c), !!! chks)

f(3, 2, 1)
\dontrun{
f(3, 1, 2)}

}
