% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R
\name{vld_err_msg}
\alias{vld_err_msg}
\alias{vld_err_msg<-}
\alias{new_err_msg}
\title{Get or set a validation error message}
\usage{
vld_err_msg(f)

vld_err_msg(f, env = parent.frame()) <- value

new_err_msg(msg, env = parent.frame())
}
\arguments{
\item{f}{Predicate function.}

\item{env}{Environment that is in scope when the \code{\{\{}-enclosed substrings
of the error message are interpolated.}

\item{value}{Error message (string or \link[rlang:quosure]{quosure} of a
string).}

\item{msg}{Error message (string).}
}
\value{
Quosure of a string.
}
\description{
Get or set a validation error message
}
\examples{
is_integer <- is.integer
vld_err_msg(is_integer) <- "{{.}} not of integer type (type: {typeof(.)})"
vld_err_msg(is_integer)

foo <- firmly(identity, is_integer)
foo(1:3)
\dontrun{
foo(runif(1:3))}

is_integer <- rlang::is_integer
msg <- local({
  len <- function(n) if (is.null(n)) "" else paste(" of length", n)
  new_err_msg("{{.}} is not an integer vector{{len(.value$n)}}")
})
vld_err_msg(is_integer) <- msg

foo <- firmly(identity, is_integer(n = 3))
foo(1:3)
\dontrun{
foo(1:2)}

}
