% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors.R
\name{vld_exprs}
\alias{vld_exprs}
\title{Specify validation expressions}
\usage{
vld_exprs(...)
}
\arguments{
\item{...}{Expressions to validate.}
}
\description{
Specify validation expressions
}
\examples{
f <- function(x, y) "Pass"

## Make a positivity checker
chk_exprs <- vld_exprs(x, x - y)
foo <- firmly(f, "{{.}} is not positive" := {isTRUE(. > 0)}(!!! chk_exprs))

foo(2, 1)
\dontrun{
foo(1, 2)}

}
\seealso{
\code{\link[=vld_checks]{vld_checks()}}
}
