% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_first_true}
\alias{which_first_true}
\title{Find first predicate to evaluate to \code{TRUE}}
\usage{
which_first_true(predicates, x, .no_true = NULL)
}
\arguments{
\item{predicates}{List of predicate functions.}

\item{x}{R object to test.}

\item{.no_true}{Value to return}
}
\value{
Index of first predicate in \code{predicates} to evaluate \code{TRUE}
  on \code{x}, otherwise \code{.no_true}.
}
\description{
Find first predicate to evaluate to \code{TRUE}
}
\examples{
predicates <- list(is.null, function(.) is.logical(.) && is.na(.))
which_first_true(predicates, NULL)  # 1
which_first_true(predicates, NA)    # 2
which_first_true(predicates, 0)     # NULL
which_first_true(predicates, sin)   # NULL
}
\keyword{internal}

