% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{with_sig}
\alias{with_sig}
\title{Rewrite the argument signature of a function}
\usage{
with_sig(.f, .sig, .attrs)
}
\arguments{
\item{.f}{Function.}

\item{.sig}{Pairlist.}

\item{.attrs}{Named list (of attributes).}
}
\value{
Function.
}
\description{
Rewrite the argument signature of a function
}
\keyword{internal}
