% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{checker-pattern}
\alias{checker-pattern}
\title{Pattern checkers}
\arguments{
\item{\dots}{Expressions to validate.}

\item{pattern}{Regular expression.}

\item{ignore.case}{Should pattern matching ignore case?}

\item{perl}{Should Perl-compatible regular expressions be used?}
}
\description{
Pattern checkers
}
\examples{
ymd <- function(y, m, d) paste(y, m, d, sep = "/")
too_old <- "Not a 21st-century year"
recent_ymd <- firmly(ymd, too_old := vld_grepl("^20[[:digit:]]{2}$", y))

recent_ymd(2017, 01, 01)
\dontrun{
recent_ymd(1999, 01, 01)}

}
\seealso{
\code{\link[=grepl]{grepl()}}
}
\alias{vld_grepl}
\usage{
vld_grepl(pattern, ..., ignore.case = FALSE, perl = FALSE)
}
