% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{checker-sets}
\alias{checker-sets}
\title{Set comparison checkers}
\arguments{
\item{\dots}{Expressions to validate}

\item{set}{Reference set (as a vector).}
}
\description{
Set comparison checkers
}
\alias{vld_in}
\alias{vld_not_in}
\alias{vld_include}
\alias{vld_exclude}
\alias{vld_within}
\alias{vld_intersect}
\alias{vld_avoid}
\alias{vld_setequal}
\usage{
vld_in(set, ...)

vld_not_in(set, ...)

vld_include(set, ...)

vld_exclude(set, ...)

vld_within(set, ...)

vld_intersect(set, ...)

vld_avoid(set, ...)

vld_setequal(set, ...)
}
