% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist.R
\name{checklist}
\alias{checklist}
\alias{is_check_formula}
\alias{is_checklist}
\title{Is a formula a check formula?}
\usage{
is_checklist(x)

is_check_formula(x)
}
\arguments{
\item{x}{R object.}
}
\value{
\code{is_check_formula(x)}, resp. \code{is_checklist(x)}, returns
  \code{TRUE} or \code{FALSE}, according to whether \code{x} is or is not a
  check formula, resp. checklist.
}
\description{
\code{is_check_formula(x)} checks whether \code{x} is a check formula, while
\code{is_checklist(x)} checks whether \code{x} is a \emph{checklist}, i.e., a
list of check formulae. Neither function verifies logical consistency.
}
\details{
If \code{is_check_formula()} cannot be evaluated on one of the
  components of a list \code{x}, then \code{is_checklist(x)} returns
  \code{FALSE}.
}
\examples{
# Valid checklist
is_checklist(list(list(~x, ~y) ~ is.numeric, "Not positive" ~ {. > 0}))

# Invalid checklists
is_checklist(list(is.numeric ~ list(~ x)))        # Backwards
is_checklist(list(list(log ~ x) ~ is.character))  # Invalid check item
}
\seealso{
\link{strictly()}, for information on the specificaton and use of
  check formulae.
}

