% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checklist}
\alias{checklist}
\alias{is_check_formula}
\alias{is_checklist}
\title{Is a checklist correctly formed?}
\usage{
is_checklist(x)

is_check_formula(x)
}
\arguments{
\item{x}{R object.}
}
\value{
\code{TRUE} or \code{FALSE}, according to whether \code{x} is a valid
  checklist.
}
\description{
A checklist is a list of \emph{check formulae} \code{q ~ p}:
\itemize{
  \item \code{p} is a predicate function, or one-sided formulaic shorthand
    thereof (see \code{purrr::\link[purrr]{as_function}()})
  \item \code{q} is a \emph{check item} for \code{p}
}
A check item is either empty, a string, or a list of formulae of the form
\code{~<expr>} or \code{<string> ~ <expr>}. (The predicate \code{p} performs
its check by being called on \code{<expr>}.)
}
\details{
\code{is_checklist()} is used to check the validity of the checks
  passed to \code{\link{strictly}()}, either via \code{...} or the argument
  \code{.checklist}.
}
\examples{
# Valid checklist
is_checklist(list(list(~x, ~y) ~ is.numeric, "Not positive" ~ ~{. > 0}))

# Invalid checklists
is_checklist(list(is.numeric ~ list(~ x)))
is_checklist(list(list(log ~ x) ~ is.character))
}

