% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promises.R
\name{lazy_assign}
\alias{lazy_assign}
\title{Assign symbols to an environment as promises}
\usage{
lazy_assign(symbs, env_eval, env_asgn)
}
\arguments{
\item{symbs}{List of symbols.}

\item{env_eval}{Common environment in which \code{symbs} are to be evaluated.}

\item{env_asgn}{Environment to which \code{symbs} are to be bound.}
}
\value{
The environment \code{env_asgn}, invisibly.
}
\description{
Assign symbols to an environment as promises
}
\keyword{internal}

