% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{misc-checkers}
\alias{misc-checkers}
\title{Miscellaneous type checkers}
\arguments{
\item{...}{Expressions to check, supplied as one-sided formulae.}
}
\value{
Check formula.
}
\description{
Miscellaneous type checkers
}
\seealso{
\code{\link[purrr]{is_empty}}, \code{\link[purrr]{is_formula}}

Other type checkers: \code{\link{bare-type-checkers}},
  \code{\link{scalar-type-checkers}},
  \code{\link{type-checkers}}
}
\alias{vld_empty}
\alias{vld_formula}
\usage{
vld_empty(...)

vld_formula(...)
}
