% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates-boolean}
\alias{predicates-boolean}
\title{Boolean predicates}
\arguments{
\item{\dots}{Expressions to validate}
}
\description{
Boolean predicates
}
\alias{vld_true}
\alias{vld_is}
\alias{vld_false}
\alias{vld_not}
\alias{vld_all}
\alias{vld_none}
\alias{vld_any}
\usage{
vld_true(...)

vld_is(...)

vld_false(...)

vld_not(...)

vld_all(...)

vld_none(...)

vld_any(...)
}
