% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates-object}
\alias{predicates-object}
\title{Object predicates}
\arguments{
\item{x}{Object to test.}
}
\description{
Object predicates
}
\examples{
row_sums <- firmly(rowSums, vld_matrix(x))
row_sums(matrix(1:6, 2, 3))

## Meaningless to sum across rows when column units differ
\dontrun{
row_sums(mtcars)}

}
\seealso{
\link[=predicates-type]{Type predicates}
}
\alias{vld_call}
\alias{vld_factor}
\alias{vld_data_frame}
\alias{vld_matrix}
\alias{vld_formula}
\alias{vld_function}
\usage{
vld_call(x)

vld_factor(x)

vld_data_frame(x)

vld_matrix(x)

vld_formula(x)

vld_function(x)
}
