% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates-object}
\alias{predicates-object}
\title{Object predicates}
\arguments{
\item{\dots}{Expressions to validate}
}
\description{
Object predicates
}
\alias{vld_call}
\alias{vld_factor}
\alias{vld_data_frame}
\alias{vld_matrix}
\alias{vld_formula}
\alias{vld_function}
\alias{vld_quosure}
\alias{vld_quosures}
\usage{
vld_call(...)

vld_factor(...)

vld_data_frame(...)

vld_matrix(...)

vld_formula(...)

vld_function(...)

vld_quosure(...)

vld_quosures(...)
}
