% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates-pattern}
\alias{predicates-pattern}
\title{Pattern predicates}
\arguments{
\item{\dots}{Expressions to validate}
}
\description{
Pattern predicates
}
\alias{vld_grepl}
\usage{
vld_grepl(pattern, ..., ignore.case = FALSE, perl = FALSE)
}
