% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates-property}
\alias{predicates-property}
\title{Property predicates}
\arguments{
\item{\dots}{Expressions to validate}
}
\description{
Property predicates
}
\alias{vld_not_null}
\alias{vld_not_empty}
\alias{vld_singleton}
\alias{vld_not_na}
\alias{vld_has_no_na}
\alias{vld_sorted}
\alias{vld_recursive}
\alias{vld_named}
\usage{
vld_not_null(...)

vld_not_empty(...)

vld_singleton(...)

vld_not_na(...)

vld_has_no_na(...)

vld_sorted(...)

vld_recursive(...)

vld_named(...)
}
