% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates-property}
\alias{predicates-property}
\title{Property predicates}
\arguments{
\item{x}{Object to test.}

\item{nm, nms}{Name(s).}

\item{n}{Length.}

\item{which}{Object attribute(s).}

\item{what}{Class name.}
}
\description{
Property predicates
}
\examples{
f <- function(x, y) NULL
foo <- firmly(f, "x, y are not disjoint" := vld_empty(intersect(x, y)))
foo(letters[1:3], letters[4:5])
\dontrun{
foo(letters[1:3], letters[3:5])}

}
\seealso{
\link[=predicates-set]{Set predicates},
\code{\link[=vld_null]{vld_null()}}, \code{\link[=vld_not_null]{vld_not_null()}}
}
\alias{vld_empty}
\alias{vld_not_empty}
\alias{vld_singleton}
\alias{vld_not_na}
\alias{vld_without_na}
\alias{vld_named}
\alias{vld_has_name}
\alias{vld_has_names}
\alias{vld_has_length}
\alias{vld_has_attr}
\alias{vld_has_attrs}
\alias{vld_inherits}
\usage{
vld_empty(x)

vld_not_empty(x)

vld_singleton(x)

vld_not_na(x)

vld_without_na(x)

vld_named(x)

vld_has_name(x, nm)

vld_has_names(x, nms)

vld_has_length(x, n)

vld_has_attr(x, which)

vld_has_attrs(x, which)

vld_inherits(x, what)
}
