% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates-relational}
\alias{predicates-relational}
\title{Relational predicates}
\arguments{
\item{x}{Object to test.}

\item{to}{Reference object.}

\item{lwr, upr}{Lower/upper bound.}

\item{na.rm}{Should \code{NA} values be disregarded?}
}
\description{
Relational predicates
}
\examples{
f <- function(x, y) log(y - x) / log(x)
foo <- firmly(f, vld_gt(0, x - 1, "y not greater than x" := y - x))
foo(2, 4)
\dontrun{
foo(1, 2)
foo(2, 2)}

}
\seealso{
\code{\link[=all.equal]{all.equal()}}, \code{\link[=identical]{identical()}}
}
\alias{vld_identical}
\alias{vld_not_identical}
\alias{vld_equal}
\alias{vld_not_equal}
\alias{vld_equivalent}
\alias{vld_not_equivalent}
\alias{vld_gt}
\alias{vld_lt}
\alias{vld_gte}
\alias{vld_lte}
\usage{
vld_identical(x, to)

vld_not_identical(x, to)

vld_equal(x, to)

vld_not_equal(x, to)

vld_equivalent(x, to)

vld_not_equivalent(x, to)

vld_gt(x, lwr, na.rm = FALSE)

vld_lt(x, upr, na.rm = FALSE)

vld_gte(x, lwr, na.rm = FALSE)

vld_lte(x, upr, na.rm = FALSE)
}
