% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates-scalar-type}
\alias{predicates-scalar-type}
\title{Scalar type predicates}
\arguments{
\item{\dots}{Expressions to validate}
}
\description{
Scalar type predicates
}
\alias{vld_number}
\alias{vld_whole_number}
\alias{vld_boolean}
\alias{vld_string}
\usage{
vld_number(...)

vld_whole_number(...)

vld_boolean(...)

vld_string(...)
}
