% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates-set}
\alias{predicates-set}
\title{Set predicates}
\arguments{
\item{x}{Object to test.}

\item{set}{Reference set (as a vector).}
}
\description{
Set predicates
}
\examples{
s3methods <- function(x) {
  methods <- attr(methods(class = class(x)), "info")
  with(methods, generic[!isS4])
}
foo <- fasten(
  chk_include("predict", s3methods(object))
)(
  function(object, data) {
    pred <- predict(object, data)
    "Do something with prediction"
  }
)

mdl <- lm(mpg ~ wt, mtcars[1:10, ])
data <- mtcars[11:12, ]

foo(mdl, data)
\dontrun{
foo(NULL, data)}

}
\seealso{
\link[=setequal]{Set operations},
\link[=predicates-property]{Property predicates}
}
\alias{chk_in}
\alias{chk_not_in}
\alias{chk_include}
\alias{chk_exclude}
\alias{chk_within}
\alias{chk_intersect}
\alias{chk_avoid}
\alias{chk_setequal}
\usage{
chk_in(x, set)

chk_not_in(x, set)

chk_include(x, set)

chk_exclude(x, set)

chk_within(x, set)

chk_intersect(x, set)

chk_avoid(x, set)

chk_setequal(x, set)
}
