% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates-type}
\alias{predicates-type}
\title{Type predicates}
\arguments{
\item{\dots}{Expressions to validate}
}
\description{
Type predicates
}
\alias{vld_closure}
\alias{vld_numerical}
\alias{vld_null}
\alias{vld_symbol}
\alias{vld_pairlist}
\alias{vld_environment}
\alias{vld_atomic}
\alias{vld_list}
\alias{vld_vector}
\alias{vld_logical}
\alias{vld_integer}
\alias{vld_double}
\alias{vld_character}
\alias{vld_raw}
\usage{
vld_closure(...)

vld_numerical(..., n = NULL)

vld_null(...)

vld_symbol(...)

vld_pairlist(...)

vld_environment(...)

vld_atomic(..., n = NULL)

vld_list(..., n = NULL)

vld_vector(..., n = NULL)

vld_logical(..., n = NULL)

vld_integer(..., n = NULL)

vld_double(..., n = NULL)

vld_character(..., n = NULL, encoding = NULL)

vld_raw(..., n = NULL)
}
