% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates}
\alias{predicates}
\title{Predicates}
\arguments{
\item{\dots}{Expressions to check.}

\item{.ref}{Reference object.}
}
\description{
Predicates
}
\alias{vld_true}
\alias{vld_is}
\alias{vld_false}
\alias{vld_not}
\alias{vld_all}
\alias{vld_none}
\alias{vld_any}
\alias{vld_closure}
\alias{vld_numerical}
\alias{vld_null}
\alias{vld_symbol}
\alias{vld_pairlist}
\alias{vld_environment}
\alias{vld_language}
\alias{vld_list}
\alias{vld_logical}
\alias{vld_integer}
\alias{vld_double}
\alias{vld_complex}
\alias{vld_character}
\alias{vld_raw}
\alias{vld_scalar_numerical}
\alias{vld_number}
\alias{vld_whole_number}
\alias{vld_boolean}
\alias{vld_string}
\alias{vld_scalar_atomic}
\alias{vld_scalar_list}
\alias{vld_scalar_logical}
\alias{vld_scalar_integer}
\alias{vld_scalar_double}
\alias{vld_scalar_complex}
\alias{vld_scalar_character}
\alias{vld_scalar_raw}
\alias{vld_scalar_vector}
\alias{vld_call}
\alias{vld_factor}
\alias{vld_data_frame}
\alias{vld_matrix}
\alias{vld_formula}
\alias{vld_function}
\alias{vld_quosure}
\alias{vld_quosures}
\alias{vld_not_null}
\alias{vld_not_empty}
\alias{vld_singleton}
\alias{vld_not_na}
\alias{vld_has_no_na}
\alias{vld_sorted}
\alias{vld_recursive}
\alias{vld_named}
\alias{vld_all_map}
\alias{vld_any_map}
\alias{vld_grepl}
\alias{vld_has_name}
\alias{vld_has_names}
\alias{vld_has_length}
\alias{vld_has_attr}
\alias{vld_has_attrs}
\alias{vld_inherits}
\alias{vld_identical}
\alias{vld_not_identical}
\alias{vld_equal}
\alias{vld_not_equal}
\alias{vld_all_equal}
\alias{vld_not_all_equal}
\alias{vld_gt}
\alias{vld_lt}
\alias{vld_gte}
\alias{vld_lte}
\alias{vld_in}
\alias{vld_not_in}
\alias{vld_contains}
\alias{vld_doesnt_contain}
\alias{vld_contained_in}
\alias{vld_not_contained_in}
\alias{vld_intersects}
\alias{vld_doesnt_intersect}
\alias{vld_setequal}
\usage{
vld_true(...)

vld_is(...)

vld_false(...)

vld_not(...)

vld_all(...)

vld_none(...)

vld_any(...)

vld_closure(...)

vld_numerical(...)

vld_null(...)

vld_symbol(...)

vld_pairlist(...)

vld_environment(...)

vld_language(...)

vld_list(...)

vld_logical(...)

vld_integer(...)

vld_double(...)

vld_complex(...)

vld_character(...)

vld_raw(...)

vld_scalar_numerical(...)

vld_number(...)

vld_whole_number(...)

vld_boolean(...)

vld_string(...)

vld_scalar_atomic(...)

vld_scalar_list(...)

vld_scalar_logical(...)

vld_scalar_integer(...)

vld_scalar_double(...)

vld_scalar_complex(...)

vld_scalar_character(...)

vld_scalar_raw(...)

vld_scalar_vector(...)

vld_call(...)

vld_factor(...)

vld_data_frame(...)

vld_matrix(...)

vld_formula(...)

vld_function(...)

vld_quosure(...)

vld_quosures(...)

vld_not_null(...)

vld_not_empty(...)

vld_singleton(...)

vld_not_na(...)

vld_has_no_na(...)

vld_sorted(...)

vld_recursive(...)

vld_named(...)

vld_all_map(.ref, ...)

vld_any_map(.ref, ...)

vld_grepl(.ref, ...)

vld_has_name(.ref, ...)

vld_has_names(.ref, ...)

vld_has_length(.ref, ...)

vld_has_attr(.ref, ...)

vld_has_attrs(.ref, ...)

vld_inherits(.ref, ...)

vld_identical(.ref, ...)

vld_not_identical(.ref, ...)

vld_equal(.ref, ...)

vld_not_equal(.ref, ...)

vld_all_equal(.ref, ...)

vld_not_all_equal(.ref, ...)

vld_gt(.ref, ...)

vld_lt(.ref, ...)

vld_gte(.ref, ...)

vld_lte(.ref, ...)

vld_in(.ref, ...)

vld_not_in(.ref, ...)

vld_contains(.ref, ...)

vld_doesnt_contain(.ref, ...)

vld_contained_in(.ref, ...)

vld_not_contained_in(.ref, ...)

vld_intersects(.ref, ...)

vld_doesnt_intersect(.ref, ...)

vld_setequal(.ref, ...)
}
