% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{properties}
\alias{is_strict_closure}
\alias{properties}
\alias{stc_args}
\alias{stc_checks}
\alias{stc_core}
\title{Properties of a strict closure}
\usage{
is_strict_closure(x)

stc_core(..f)

stc_checks(..f)

stc_args(..f)
}
\arguments{
\item{x}{R object.}

\item{..f}{Strict closure, i.e., interpreted function of class
\code{"strict_closure"}.}
}
\description{
Properties of a strict closure
}

