% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope.R
\name{scope}
\alias{globalize_check}
\alias{localize_check}
\alias{scope}
\title{Convert the scope of a check formula}
\usage{
localize_check(chk)

globalize_check(chkr)
}
\arguments{
\item{chk}{Check formula of global scope with a custom error message, i.e., a
formula of the form \code{<string> ~ <predicate>}, cf. \link{strictly}.}

\item{chkr}{Function of class \code{"local_checker"}, i.e., a function
created by \code{localize_check()}.}
}
\description{
\code{localize_check()} converts a check formula of global scope into a
function that generates a check formulae of local scope;
\code{globalize_check()} takes such a function and returns the underlying
check formula (of global scope). These operations are mutually invertible.
}
\examples{
\dontrun{

is_positive <- localize_check("Not positive" ~ {. > 0})
is_positive(x, x - y)
#> list("Not positive: `x`" ~ x, "Not positive: `x - y`" ~ x - y) ~ {. > 0}

globalize_check(is_positive)
#> "Not positive" ~ {. > 0}
}

}
\seealso{
\link{strictly} explains the notion of "scope" for check formulae.
}

