% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{vld}
\alias{vld}
\title{Capture input-validation checks}
\usage{
vld(...)
}
\arguments{
\item{...}{Input-validation checks as error message-check definitions (cf.
\code{\link[rlang:dots_definitions]{rlang::dots_definitions()}}).}
}
\value{
List of pairs of quosures, with named components \code{msg} (error
message) and \code{chk} (input validation check), of class \code{validationChecks}.
}
\description{
Capture input-validation checks
}
\examples{
vld(!!!vld(!!!vld(f, a := b, c := d)), g, e := f)

}
