% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{vld}
\alias{vld}
\title{Capture input validation checks}
\usage{
vld(...)
}
\arguments{
\item{...}{Input-validation checks as error message-check definitions (cf.
\code{\link[rlang:dots_definitions]{rlang::dots_definitions()}}).}
}
\value{
List of pairs of quosures, with named components \code{msg} (error
message) and \code{chk} (input validation check), of class \code{validation_checks}.
}
\description{
Capture input validation checks
}
\details{
The order of the input validation checks is not necessarily
preserved.
}
\examples{
# Stipulate that a, b, c are numeric and a > b > c
z <- 0
vld("Not numeric" := is.numeric,
    !!! vld({isTRUE(. > !! z)} ~ vld(a - b, b - c)))

}
