% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors.R
\name{vld_checks}
\alias{vld_checks}
\alias{vld_exprs}
\title{Specify validation checks}
\usage{
vld_checks(...)

vld_exprs(...)
}
\arguments{
\item{...}{Checks or expressions to check.}
}
\description{
Specify validation checks
}
